Use vscode-js-debug instead of deprecated node-debug2
This also restore ability to debug running Node processes (eg start with
--inspect).

Requires usage of latest LSP4E support for DAP.

Fixes https://github.com/eclipse-wildwebdeveloper/wildwebdeveloper/issues/555
