+++ org.eclipse.wildwebdeveloper.tests/testProjects/HelloWorldTS/.build/index.js
@@ -1,4 +1,4 @@
 "use strict";
-var user = "Eclipse User";
-console.log("Hello world,, " + user + '!');
+let user = "Eclipse User";
+new Promise(resolve => setTimeout(resolve, 1000)).then(() => console.log("Hello world,, " + user + '!'));
 //# sourceMappingURL=index.js.map
