+++ org.eclipse.wildwebdeveloper.tests/testProjects/HelloWorldTS/index.ts
@@ -1,2 +1,4 @@
 let user:string = "Eclipse User";
-console.log("Hello world,, " + user  + '!');
+new Promise(resolve => setTimeout(resolve, 1000)).then(() => 
+    console.log("Hello world,, " + user  + '!')
+);
