+++ org.eclipse.wildwebdeveloper.tests/testProjects/HelloWorldTS/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "es2017",
     "module": "commonjs",
     "lib": ["es6", "dom"],
     "allowJs": true,
@@ -8,7 +8,6 @@
     "outDir": ".build/",
     "strict": true,
     "noImplicitAny": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "skipLibCheck": true
   }
 }
