+++ org.eclipse.wildwebdeveloper/META-INF/MANIFEST.MF
@@ -21,7 +21,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.resources;bundle-version="3.12.0",
  org.eclipse.debug.core;bundle-version="3.11.0",
  org.eclipse.debug.ui;bundle-version="3.11.0",
- org.eclipse.lsp4e.debug;bundle-version="0.9.0",
+ org.eclipse.lsp4e.debug;bundle-version="0.15.5",
  com.google.gson;bundle-version="2.8.2",
  org.eclipse.core.expressions;bundle-version="3.6.0",
  org.eclipse.tm4e.languageconfiguration,
