+++ org.eclipse.wildwebdeveloper/build.properties
@@ -10,6 +10,7 @@ bin.includes = META-INF/,\
                snippets/,\
                icons/,\
                schema/,\
+               js-debug/,\
                plugin.properties
 # See https://github.com/redhat-developer/yaml-language-server/issues/253
 bin.excludes = node_modules/yaml-language-server/out/server/node_modules/
