+++ org.eclipse.wildwebdeveloper/package.json
@@ -15,7 +15,6 @@
     "vscode-json-languageserver": "file:target/vscode-json-languageserver-1.3.4.tgz",
     "debugger-for-chrome": "file:target/debugger-for-chrome-4.13.0.tgz",
     "eslint-server": "file:target/eslint-server-2.4.1.tgz",
-    "node-debug2": "file:target/node-debug2-1.43.0.tgz",
     "@vue/language-server" : "1.8.15",
     "fsevents" : "2.3.3"
   }
