+++ target-platform/target-platform.target
@@ -41,7 +41,9 @@
 	<unit id="org.eclipse.mylyn.wikitext_feature.feature.group" version="0.0.0"/>
 </location>
 <location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
-	<repository location="https://download.eclipse.org/lsp4e/releases/latest/"/>
+	<!--⚠️ Use release when LSP4E release contains https://github.com/eclipse/lsp4e/commit/909c41fb6d80895388634f152b7f9dd20f9482b8 https://github.com/eclipse/lsp4e/commit/909c41fb6d80895388634f152b7f9dd20f9482b8 -->
+	<!--repository location="https://download.eclipse.org/lsp4e/releases/latest/"/-->
+	<repository location="https://download.eclipse.org/lsp4e/snapshots/"/>
 	<unit id="org.eclipse.lsp4e" version="0.0.0"/>
 	<unit id="org.eclipse.lsp4e.debug" version="0.0.0"/>
 	<unit id="org.eclipse.lsp4j" version="0.0.0"/>
